# public/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Jubilujah Internal API",
    "version": "1.0.0",
    "summary": "Editorial, pipeline, ratings, awards, radio, and admin surface for Jubilujah.",
    "description": "OpenAPI 3.1 contract implementing every endpoint enumerated in Jubilujah Build Spec v1.0 §19 (API Surface — Initial). Authentication is delegated to JubileeInspire SSO (Build Spec §5); the resulting session cookie (`session_id`, HttpOnly/Secure/SameSite=Strict) authorizes all calls. All mutating requests require a CSRF token header (`X-CSRF-Token`). Role-based gates per §5: minimum `content_editor` for read; `production_manager` for pipeline transitions; `radio_producer` for playlist mutations; `admin` for `/admin/*`. Domain schemas live at `/schemas/v1/*.schema.json` and are referenced via `$ref` from this document.",
    "contact": {
      "name": "Gabriel Ungureanu — Stewardship",
      "url": "https://jubilujah.com"
    },
    "license": {
      "name": "Proprietary — Jubilee Music Internal Use Only",
      "identifier": "LicenseRef-Jubilee-Internal"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3119/api",
      "description": "Local development (server.js)"
    },
    {
      "url": "https://jubilujah.com/api",
      "description": "Production stub (gated by JubileeInspire SSO)"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    }
  ],
...
  "paths": {
    "/artists": {
      "get": {
        "tags": ["Catalog"],
        "operationId": "listArtists",
        "summary": "List artists, optionally filtered by grouping.",
        "parameters": [
          { "$ref": "#/components/parameters/GroupingQuery" },
          { "$ref": "#/components/parameters/LimitQuery" },
          { "$ref": "#/components/parameters/OffsetQuery" }
        ],
        "responses": {
          "200": {
            "description": "Page of artists.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ArtistList" },
                "examples": {
                  "inspireFamily": { "$ref": "#/components/examples/ArtistListInspireFamily" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "post": {
        "tags": ["Catalog"],
        "operationId": "createArtist",
        "summary": "Create a new artist (content_editor+).",
        "parameters": [{ "$ref": "#/components/parameters/CsrfHeader" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ArtistCreate" },
              "examples": {
                "affiliated": { "$ref": "#/components/examples/ArtistCreateAffiliated" }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Artist created.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Artist" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" }
        }
      }
    },
    "/artists/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
      "get": {
        "tags": ["Catalog"],
        "operationId": "getArtist",
        "summary": "Fetch a single artist by id.",
        "responses": {
          "200": {
            "description": "Artist found.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Artist" },
                "examples": {
                  "melodyInspire": { "$ref": "#/components/examples/ArtistMelody" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "tags": ["Catalog"],
        "operationId": "updateArtist",
        "summary": "Partial update of an artist record.",
        "parameters": [{ "$ref": "#/components/parameters/CsrfHeader" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ArtistPatch" } }
          }
        },
        "responses": {
          "200": {
            "description": "Updated artist.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Artist" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" }
        }
      }
    },
...
    "/albums/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
      "get": {
        "tags": ["Catalog"],
        "operationId": "getAlbum",
        "summary": "Fetch a single album.",
        "responses": {
          "200": {
            "description": "Album found.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Album" },
                "examples": {
                  "breakthrough": { "$ref": "#/components/examples/AlbumBreakthrough" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "tags": ["Catalog"],
        "operationId": "updateAlbum",
        "summary": "Partial update of an album.",
        "parameters": [{ "$ref": "#/components/parameters/CsrfHeader" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/AlbumPatch" } }
          }
        },
        "responses": {
          "200": {
            "description": "Updated album.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Album" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session_id",
        "description": "HttpOnly + Secure + SameSite=Strict cookie issued by the backend after a successful JubileeInspire SSO exchange (Build Spec §5)."
      }
    },
    "parameters": {
      "IdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "UUID of the resource.",
        "schema": { "type": "string", "format": "uuid" }
      },
...
      "CsrfHeader": {
        "name": "X-CSRF-Token",
        "in": "header",
        "required": true,
        "description": "Double-submit CSRF token issued at login (Build Spec §20).",
        "schema": { "type": "string", "minLength": 16 }
      },
      "GroupingQuery": {
        "name": "grouping",
        "in": "query",
        "required": false,
        "description": "Restrict to one artist grouping (Build Spec §6).",
        "schema": {
          "type": "string",
          "enum": ["inspire_family", "affiliated", "childrens_brands", "other_initiatives"]
        }
      },
...
      "LimitQuery": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Page size (default 50, max 200).",
        "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 50 }
      },
      "OffsetQuery": {
        "name": "offset",
        "in": "query",
        "required": false,
        "description": "Pagination offset.",
        "schema": { "type": "integer", "minimum": 0, "default": 0 }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "No valid session cookie present.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "examples": {
              "missingSession": {
                "value": { "error": "unauthorized", "message": "Session required. Re-authenticate via JubileeInspire SSO." }
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated but role does not satisfy the route's minimum.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "examples": {
              "needsAdmin": {
                "value": { "error": "forbidden", "message": "Role 'admin' required for this operation." }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource does not exist or is not visible to the caller.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "examples": {
              "missing": { "value": { "error": "not_found", "message": "Resource not found." } }
            }
          }
        }
      },
      "UnprocessableEntity": {
        "description": "Request body failed Pydantic / domain validation.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ValidationError" },
            "examples": {
              "hebrewArticle": {
                "value": {
                  "error": "validation_failed",
                  "message": "Hebrew article rule violation in lyrics.",
                  "details": [
                    { "field": "lyrics.text", "code": "hebrew_article_rule", "hint": "Use 'the Ruach Kodesh' or 'Ruach HaKodesh' — never 'the Ruach HaKodesh'." }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
      "Error": { "$ref": "/schemas/v1/error.schema.json" },
      "ValidationError": { "$ref": "/schemas/v1/validation-error.schema.json" },
      "Artist":         { "$ref": "/schemas/v1/artist.schema.json" },
      "ArtistCreate":   { "$ref": "/schemas/v1/artist.create.schema.json" },
      "ArtistPatch":    { "$ref": "/schemas/v1/artist.patch.schema.json" },
      "ArtistList":     { "$ref": "/schemas/v1/artist.list.schema.json" },
      "Album":          { "$ref": "/schemas/v1/album.schema.json" },
...
      "AlbumPatch":     { "$ref": "/schemas/v1/album.patch.schema.json" },
...
    },
    "examples": {
      "ArtistListInspireFamily": {
        "summary": "First page of Inspire Family artists.",
        "value": {
          "items": [
            {
              "id": "0c0a1d2e-1111-4aaa-9000-000000000001",
              "display_name": "Melody Inspire",
              "slug": "melody-inspire",
              "grouping": "inspire_family",
              "five_fold": ["pastor", "evangelist"],
              "genre_anchor": "contemporary-worship",
              "ohi_default": true
            },
            {
              "id": "0c0a1d2e-1111-4aaa-9000-000000000006",
              "display_name": "Caleb Inspire",
              "slug": "caleb-inspire",
              "grouping": "inspire_family",
              "five_fold": ["prophet", "teacher"],
              "genre_anchor": "modern-gospel-rock",
              "ohi_default": true
            }
          ],
          "limit": 50,
          "offset": 0,
          "total": 13
        }
      },
      "ArtistCreateAffiliated": {
        "summary": "Create an affiliated artist record.",
        "value": {
          "display_name": "Hosanna Collective",
          "slug": "hosanna-collective",
          "grouping": "affiliated",
          "genre_anchor": "modern-worship",
          "bio": "Multi-vocalist worship collective affiliated with Jubilee Music."
        }
      },
      "ArtistMelody": {
        "summary": "Single artist response for Melody Inspire.",
        "value": {
          "id": "0c0a1d2e-1111-4aaa-9000-000000000001",
          "display_name": "Melody Inspire",
          "slug": "melody-inspire",
          "grouping": "inspire_family",
          "five_fold": ["pastor", "evangelist"],
          "genre_anchor": "contemporary-worship",
          "ohi_default": true,
          "bio": "Lead vocalist of the Inspire Family; pastoral-evangelistic anchor.",
          "created_at": "2025-01-15T10:00:00Z",
          "updated_at": "2026-05-30T18:42:11Z"
        }
      },
      "AlbumBreakthrough": {
        "summary": "Album record for Caleb Inspire's Breakthrough.",
        "value": {
          "id": "1a2b3c4d-2222-4bbb-9000-000000000010",
          "artist_id": "0c0a1d2e-1111-4aaa-9000-000000000006",
          "title": "Breakthrough",
          "slug": "breakthrough",
          "release_date": "2026-04-12",
          "language": "en",
          "song_count": 12,
          "twelve_song_lock_satisfied": true,
          "ohi_label": true
        }
      },
...
    }
  }
}